fix: choose queue-based approach for embeddings
